ws.go: Fix first race condition in cancelPool management
Cancel occurs in async. When the cancelled function exits, it deletes
the entry in cancelPool, and could replace the new routine's cancel
function, which is of course undesirable. Therefore this commit causes
handleConn to stop deleting from cancelPool entirely.

Note that this introduces a memory leak at the size of
context.CancelFunc per user, because we are never deleting from
cancelPool.

This does not actually solve the race condition documented in
fc911928 where two new connections start sufficiently close to each
other.