{courses,ws}.go: Populate userCourseGroups for each connection
It is way too expensive to query the database every time we need to
check whether a user has chosen a course in a group. Since we can
(hopefully) guarantee that there is only one usable connection for any
given moment and user, we could store this data along with the
connection as a local variable in handleConn, which would be eligible
for garbage collection when handleConn exits. Here we create the
data structures that globally represent courseTypes and courseGroups,
and during the initial stages of handleConn, perform database queries to
populate userCourseGroups with the groups that the user has already
chosen.

Note that the HELLO command handler does similar database queries, and
as per the TODO listed in the comments, should be moved up here for
efficiency. (HELLO also serves as some sort of an initial connection
check; this should probably be replaced with WebSocket's native pings.)