wsc.go: Check newCtx.Done() when entering other event cases
We select the context done channel when entering other cases too (see
below) because we need to make sure the context cancel works even if
both the cancel signal and another event arrive while processing a
select cycle.

It's a bit verbose and looks repetitive but is technically more correct
than code without it.